@@ -4,22 +4,28 @@
   "private": true,
   "author": "Taiwo Damilola",
   "dependencies": {
+    "@craco/craco": "^7.1.0",
     "@material-ui/core": "^4.11.4",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/pickers": "^3.3.10",
-    "@react-pdf/renderer": "^3.1.12",
+    "@material-ui/styles": "^4.11.5",
+    "@react-pdf/renderer": "^3.4.4",
     "@reduxjs/toolkit": "^1.8.5",
     "@tanstack/react-query": "^5.51.18",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
+    "@testing-library/jest-dom": "^5.16.3",
+    "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^12.1.10",
     "aos": "^2.3.4",
+    "assert": "^2.1.0",
     "axios": "^0.27.2",
+    "browserify-zlib": "^0.2.0",
+    "buffer": "^6.0.3",
     "formik": "^2.2.9",
     "framer-motion": "^4.1.17",
     "patch-package": "^7.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "process": "^0.11.10",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-export-table-to-excel": "^1.0.6",
     "react-icons": "^4.12.0",
     "react-redux": "^8.0.2",
@@ -27,7 +33,7 @@
     "react-responsive-carousel": "^3.2.18",
     "react-responsive-masonry": "^2.1.7",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react-scripts": "5.0.1",
     "react-spinners": "^0.13.4",
     "react-to-pdf": "^0.0.14",
     "react-to-print": "^2.14.7",
@@ -35,15 +41,18 @@
     "redux": "^4.2.0",
     "redux-persist": "^6.0.0",
     "redux-thunk": "^2.4.1",
+    "safer-buffer": "^2.1.2",
+    "stream-browserify": "^3.0.0",
     "styled-components": "^5.3.0",
+    "util": "^0.12.5",
     "web-vitals": "^1.0.1",
     "xlsx": "https://cdn.sheetjs.com/xlsx-0.20.1/xlsx-0.20.1.tgz",
     "yup": "^0.32.11"
   },
   "scripts": {
-    "start": "react-scripts --openssl-legacy-provider start",
-    "build": "react-scripts --openssl-legacy-provider build",
-    "test": "react-scripts test",
+   "start": "craco start",
+    "build": "craco build",
+    "test": "craco test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
